requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
coveralls==1.5.0
docopt==0.6.2
flasgger==0.9.0
Flask==1.0.2
Flask-API==1.0
Flask-RESTful==0.3.6
flask-restplus==0.11.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
nose==1.3.7
pluggy==0.7.1
py==1.6.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.7.4
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
waitress==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
